LO: more friendly error msg for email duplicate.
When an openid user create/update profile in Moztrap,
it is possible the user has duplicated email address with
another user registered as Moztrap native user or Persona.

We can't assign the original email address directly to the
openiduser because OpenID doesn't validate the email owner.

So for now, we block such a use case for security reasons.